Fixed get_running_tasks() returning tasks from other event loops
On the trio and curio backends, the implementation used the garbage collector interface.
This ignored the possibility of multiple event loops running in the process.
The new implementation uses "native" means of obtaining the list of tasks.

It also means that it had to be changed into a coroutine function because of curio.