Gracefully handle socket creation failure
It is possible that the call to create a regular socket fails, even
before you try to do anything with it. In those cases, we should allow
another attempt to succeed.

One way that this can happen is on linux with `ipv6.disable=1` on the
kernel command line. In this case, opening a socket with an IPv6
address family will fail with EAFNOSUPPORT.