Fixed uvloop loops leaking to regular asyncio tests
Previously, when uvloop was installed, it would get set as the event loop policy and was never replaced with the default policy, contrary to how the tests were supposed to be run. This change makes sure that the default event loop policy is restored when uvloop should not be in use.