Shutdown the loop's default thread executor at the end of a test (#504)
The same approach is adopted in IsolatedAsyncioTestCase. It prevents
leaking asyncio threads as loop.close() doesn't join them.

BaseEventLoop.shutdown_default_executor was introduced in Python 3.9.
For earlier versions, this commit introduces an adapted version of
the method: _shutdown_default_executor. However, it doesn't work
with uvloop < 0.15.0 because, uvloop doesn't expose
_default_executor attribute. It's rather a minor issue as
uvloop 0.15.0 is 2 years old - in such case this commit is a no-op.

Fixes #503.