Adding mypy argument to ignore comment error
The `mypy` hook in pre-commit fails with `error: Unused "type: ignore" comment` on several test files. Assuming that these comments are necessary, this flag configures mypy to ignore these errors and allows the hook to pass. Alternatively, the offending comments could be removed if that solution is acceptable/preferable; doing so would also allow the hook to pass.