Fix(asyncio): to_thread.async_run holding ref of Context
If the context had some objects,
after to_thread.async run,
the worker thread will keep one ref of the context,
which will cause the objects can not be released.

An example test is added as well.

This commit only fix the asyncio backend.
It should be checked for other backends.