Always set _closed = True on asyncio UDP aclose
This ensures that a closed UDPSocket will not raise BrokenResourceError.
I'm not sure whether this change is strictly necessary---it may be
impossible for transport.is_closing to return True on UDP prior to
AnyIO's aclose being called, in which case this change does nothing.
Still, I see no reason not to make this change to be safe.